Created new ts file to make test file cleaner
